Refactor address creation methods and add utility functions
