+++ odra-vm/src/vm/mod.rs
@@ -5,5 +5,6 @@ mod balance;
 mod odra_vm;
 mod odra_vm_state;
 mod storage;
+mod utils;
 
 pub use self::odra_vm::OdraVm;
